Extract copyToClipboard utility and fix CopyableCode for HTTP contexts
Centralises the execCommand fallback in a shared lib/clipboard.ts so
both InstructionsBlock and CopyableCode work on non-secure origins
(HTTP local network). No behaviour change on HTTPS.
