+++ frontend/src/lib/clipboard.ts
@@ -1,0 +1,27 @@
+/**
+ * Copies text to clipboard with a fallback for non-secure contexts (HTTP).
+ * navigator.clipboard is undefined on HTTP origins, so we fall back to the
+ * legacy execCommand approach.
+ */
+export function copyToClipboard(text: string): Promise<void> {
+  if (navigator.clipboard) {
+    return navigator.clipboard.writeText(text);
+  }
+
+  return new Promise((resolve, reject) => {
+    try {
+      const textarea = document.createElement("textarea");
+      textarea.value = text;
+      textarea.style.position = "fixed";
+      textarea.style.opacity = "0";
+      document.body.appendChild(textarea);
+      textarea.focus();
+      textarea.select();
+      document.execCommand("copy");
+      document.body.removeChild(textarea);
+      resolve();
+    } catch (err) {
+      reject(err);
+    }
+  });
+}
