+++ frontend/src/pages/dashboard/InstructionsBlock.tsx
@@ -2,6 +2,7 @@ import { useState, useRef, useEffect } from "react";
 import { Copy, Check } from "lucide-react";
 import { toast } from "sonner";
 import { Button } from "@/components/ui/button";
+import { copyToClipboard } from "@/lib/clipboard";
 
 interface InstructionsBlockProps {
   instructions: string;
@@ -33,28 +34,9 @@ export function InstructionsBlock({
       timerRef.current = setTimeout(() => setCopied(false), 2000);
     };
 
-    if (navigator.clipboard) {
-      navigator.clipboard.writeText(instructions).then(onSuccess, () =>
-        toast.error("Failed to copy to clipboard"),
-      );
-      return;
-    }
-
-    // Fallback for non-secure contexts (HTTP), e.g. local network access
-    try {
-      const textarea = document.createElement("textarea");
-      textarea.value = instructions;
-      textarea.style.position = "fixed";
-      textarea.style.opacity = "0";
-      document.body.appendChild(textarea);
-      textarea.focus();
-      textarea.select();
-      document.execCommand("copy");
-      document.body.removeChild(textarea);
-      onSuccess();
-    } catch {
-      toast.error("Failed to copy to clipboard");
-    }
+    copyToClipboard(instructions).then(onSuccess, () =>
+      toast.error("Failed to copy to clipboard"),
+    );
   }
 
   return (
